First grammar for the language
This commit includes the first explicit grammar for at least some important
parts of the language. The grammar is enough to parse the simplest Tic-Tac-Toe
example. There is also a parser program that applies the grammar.

One can run the program using Node.js:

    $ npm install && npx webpack && node build/main.js

This first grammar is by no means good or scalable to the other examples. It has
been quite a fight correctly expressing how block-type definitions can be nested
arbitrarily deep. I actually had to go and read the grammar for Python to
understand how they did it but once that happened everything clicked into place.

I'm also not sure yet if the grammar is actually usable. There are a lot of
different constructs here and I can feel that they're going to balloon into many
different semantic paths in the parsing stage. But I think it's appropriate I
try and do something with this grammar to learn what changes I might need to
make later whilst also making it work for the other examples.

I also had to update the Tic-Tac-Toe example to reflect the changes that I've
made since writing all of the other examples. Most notable is the way we do
turns – previously this was just covered using something like `many until X` but
now we tend to tie the turn to a player to ease manually faffing with a
`current_player` variable.

The grammar is written in the Ohm language. I wanted to have an
implementation-independent grammar to aid implementations in as many languages
as possible, and hopefully Ohm is well specified enough that support in other
languages will appear.