Parser to build a source tree from the grammar.
I have now leveraged the basic language grammar and automatic parser
to build a more deliberate parser. I am calling the automatic
representation the "physical source tree" and the new deliberate
representation the "abstract source tree" – because the physical tree
includes all details of the source code whereas the abstract tree
represents only the essential meaning.

This has been necessary because the automatic representation is useful
but not quite factored well enough to work with directly. The next
phase will be to resolve all of the scopes and thereby convert the
references that are currently identifier paths into actual links
between nodes. This is where the tree becomes a graph.

Resolving the scopes will involve going up the tree, which is not
something that is very clear how to achieve natively with Ohm. I'm
sure it can be done with action parameters, but would be a bit of
heavy lifting. Even then, I want to apply essential graph algorithms
to the tree (like a depth first search but up the tree instead of
down) and would prefer not to implement this myself. Graph algorithms
are not hard in principle but recoding the wheel seems foolish here.

Given that I will be using a graph data structure for the subsequent
phases, it seemed sensible to me to start with a graph library to
represent the tree (which is just a graph but with more constraints)
and then also take advantage of any built-in performant algorithms it
provides. I've chosen `graphology` for this purpose.

I am planning for there to be two more phases. The next will be the
"abstract game graph" (AGG) that will represent only the essential
logic of the game. This representation should be useful for analysis
of the game and generating metrics like complexity and weight. The
subsequent phase will be the "physical game graph" (PGG) that the
engine will actually use to run the game. This might include
optimisations of the AGG and can also be implementation dependent.
For example, I'm imagining that a database-engine-based
implementation might implement an entire subtree as a single SQL
query, in which case the PGG would only contain a single node to
represent a whole subtree of the AGG.

The first test of the code was to output visible source trees for
tic-tac-toe and cards examples. This was already fruitful: it
clarified that ranges need a special operation type of their own
because it wasn't possible to spot a range just by looking at the
tree and identifying a sequence reference with two sources. There are
also a bunch of other tweaks to bring the examples and grammar in
line.

In general the code is factored to almost completely avoid caring
about what child nodes are and only require local context to
establish the tree. I think this is a sign of a sensible grammar?
There is one niggle: deciding whether a sequence item should be
linked with a source edge or a definition edge currently requires
inspecting the child node. I think this could disappear with a better
grammar.

For now I am going to press on and try and resolve scopes. I thought
this was going to be relatively simple until I remembered about open
classes and how they add scope to pre-existing items. I haven't
managed to think through all the detail of this yet.