Locally scoped identifiers
When filtering or extending items in a sequence, the fields of each item
implicitly come into the local scope:

    x:
        a: 1
    y: x
        b: 1 + a of z

Here, `b` can refer directly to `a` rather than `a of x` because `x` is the
source sequence and its fields are implicitly in scope.

BigDeal allows the introduction of an identifier to represent these locally
scoped fields:

    x:
        a: 1
    y: many z from x
        b: 1 + a of z

In this case, the fields from the source sequence are only available through the
introduced name. This is useful where the name of a local field clashes with the
name of a field from the source sequence:

    rows of grid:
        row: 1 from row of grid
        squares: many square from grid where row of square = row

By default, the source sequence field would be unavailable because it is
shadowed by the local field, but by introducing a new name we can refer to the
source sequence fields.

The parser and scope resolver now have support for introduced local names as
part of sequence references. This was the feature that originally led us down
the rabbit hole of indentation as the test examples all use heavily indented
blocks. It also helped us uncover a bug with more than two levels of identifier
path being ordered incorrectly.