Aggregates
Bigdeal now supports parsing aggregates, linking them to the sequences they are
defined over, and resolving other identifiers to them.

## What an aggregate is

An aggregate is a sequence that is defined over another sequence but does not
just iterate and project its items. If a game makes use of `sum of x` or `count
of x` then it is using the `sum` and `count` built-in aggregates.

Aggregates are a useful meta-property of a sequence. The `grid` standard library
sequence will for example expose a `paths` aggregate that returns all of the
possible paths through squares in the grid. The `paths` aggregate is based on
the `grid` but is not just some transformation or filter over its squares – it
is a whole new sequence with different items (in this case many more).

Aggregates do not even have to source themselves from the sequence they are
defined over. They could be constant or sourced from some other sequence
entirely. A good use that appears in examples already is `secrecy` – this is an
aggregate that will almost never be sourced from the sequence it is defined over
that describes some meta-property of that sequence. So aggregates in Bigdeal are
somewhat analogous to class-level functions in other languages.

The reason to use them is that they are inheritable. If an aggregate is defined
over some source sequence, and then that sequence is filtered and transformed to
form a new sequence, then the aggregate is still applicable to the new sequence.
This makes it possible to define generic sequences like `grid` or even `number`
that define useful aggregates which still work when specialised for a specific
use case.

Note that part of the thing that makes this magic possible is the fact that
inheriting sequences in Bigdeal can only add – one cannot remove fields
from items in a sequence, only filter or add to them. So an aggregate is always
valid over derived sequences.

## Adding support for aggregates

Aggregates are created when defining a path rather than a single identifier.

    x: 1
    aggregate of x: 2

In this case, `aggregate` does not become a field of `x` but an external
sequence defined over it. Thankfully the grammar already supported multiple
identifiers in a path as part of a definition so there weren't any extra parsing
steps required.

Resolving aggregates proved to be tricky. The resolver initially tries to
resolve all of the aggregates first. This means when we come to resolve paths
the resolution algorithm can continue to assume that `x of y` first needs to
find a node named `y` and then one named `x`. Removing this assumption would
mean that the resolver would need to be able to find `y`, attempt to find `x`,
and then step back into the middle of finding `y` again – which would be doable
but significantly more complicated.

I had hoped that reusing the `defn` type edges to also refer to aggregates would
work, but alas there wasn't enough context after resolving them to be able to
work out what was an aggregate and what was a field – so I had to introduce a
new `agg` edge kind.

The fact that resolving aggregates introduces loops or multiple branches to the
same node turns the "source tree" into now a real "graph" and forced me to
confront assumptions over tree structures. For example, it is now possible to
follow two different edges to the same node that have different meanings – one
of which would allow a path to be resolved and one which would not. In this case
it's possible to visit a node once via the wrong edge and never visit it again
because it is marked as already visited.

I did think that maybe I needed to move to a model of having names live on edges
rather than on nodes, but thankfully I realised that switching to a DFS
algorithm that visited edges rather than nodes would remove the problem, and
this proved to be a simple drop-in replacement that pleasingly worked straight
away.

I also had to revisit the use of `fn` edges, which in the previous model
identified the field that was going to be applied along with a `src` edge
showing what to apply it to. I've flip-flopped on whether that model works or
whether one where the `src` edge identifies the field or aggregate and a
secondary edge identifies what it is being applied to. I've now switched to the
latter and so the resolver now generates `ctx` edges. I'm not convinced there's
much between them in reality, but for now this seems more natural.

There is also a whole algorithm to do with transforming the graph to correctly
account for inheritance that is waiting to be implemented. I've realised though
that we don't need this for now – the `ctx` edges are enough to contain all the
information. There may be a subsequent stage in the engine to remove all of the
`ctx` edges by introducing intermediate nodes that account for the inheritance.

I want to keep intermediate nodes out of the parsing stage for as long as
possible though. There is currently a useful property that every node relates
directly back to a source code range, whereas intermediate nodes would be purely
virtual. I have a feeling that a representation without intermediate nodes will
be easier to use in a future editor.

There is also some other stuff mixed in here – the greater than and less than
operations were missing, and there is now a proper Makefile which had sped up
the feedback loop on testing.

I should now be able to write a `grid` standard library module and have it be
completely parsed, so that's what I'm going to do next!