Inheritance
The parser will now pick up fields on sequences that inherit from other
sequences when resolving identifiers.

When resolving things that can be inherited (which includes both aggregates and
definitions) we need to resolve them to any *re-definition* rather than the
original definition. This is because the re-definition may have new fields which
the sequence will now include.

For example, in [this test](lang/test/scoping_agg_ctx_edge.bigdeal) the `hound`
sequence introduces a new field `happy` not on the original `dog`. Taking an
aggregate of `hound` that is defined on `dog` and iterates over `dog` should
have the `happy` field included – because it present on the more inherited
`hound`.

The way we do this is that before looking for identifiers starting from the last
node in an identifier path that we found, we now first look at all the nodes
that we have resolved so far in the identifier path and see whether any of them
represent inheritances (e.g. there is a chain of `src` edges from the earlier
node down to this one). If there is, then we will first look at any of the
`defn` edges defined on that path downwards, as these things will be in scope.

Where this involves finding a field on an inherited aggregate, as in the above
example, we need to consider `ctx` edges as well. That's because the node that
the field is defined on will not be in our `src` edges, it will only be defined
on the inherited node through a `ctx` edge.

E.g, if we are iterating over `best of hound`, when we've found `best` there
will then be a `src` edge to `best` and a `ctx` edge to `hound`. When we then
want to find `happy of best of hound`, the `happy` field is not defined on any
of the nodes reachable by `src` edges, it's defined on the `hound` node
reachable by the `ctx` edge. Hence, we need to consider `ctx` edges when finding
possible fields on inherited sequences.

To be honest, the complexity here is getting a bit much.

I'm not super happy with the implementation, because it uses `allSimplePaths` to
find all possible paths from dervied to base sequences. This feels like a hack
somehow. Something seems not right around the part where there can be multiple
paths. I don't think this code will stand the test of time – it may disintegrate
up contact with multiple source sequences derived from the same base.

I did consider an approach which used more node copying to better establish what
was in scope and when. The idea here would be that when an aggregate is
inherited, the entire tree would be copied to reference the new derived
sequence. This is not very easy to get right, and it's not very clear how it
would better resolve identifiers when copying. But maybe it could always start
resolution from the newly derived sequence, and that way automatically pick up
new identifiers?

There is still more to do here, because re-definiton of fields on derived
sequences still are not explicitly linked to their base counterparts in the
graph. Everything sort of works fine without this, but probably when we get to
building the engine this will need to be fixed. This might involve a copying
rewrite.

At the very least, this is telling me that I haven't defined the graph well
enough yet, and that I should write down somewhere a proper spec for all of
these edges and what their semantics are.