Standard library module for grids
This commit adds the long-awaited grid module to a new standard library! `grid`
has been referenced from the tic-tac-toe example from very early on, and has
popped up in many of the examples too – it turns out square grids are the basis
for most boards.

The grid here is pretty basic – it is just a generative sequence of squares
limited by a row and column number. It comes with a bunch of handy aggregates
that allow looking at the various grid squares in groups, such as in sets of
rows and columns.

What's not here yet is paths, edges or path finding, which are referenced in
later examples. This comes up a lot in games and is clearly a useful utility,
but I don't think there's any way to represent these in `bigdeal` language yet.
I have imagined there would need to be some kind of recursive computation
feature (not unlike SQL's recursion capability) that would allow running DFS
algorithms. There could also be something more specialised to just that case.

I could also of course just make that part of the code implemented in the
engine, such that it wouldn't be possible to write `grid` purely in bigdeal.
This is still an open door. Part of the reason I wanted to make `grid` and other
standard library modules pure `bigdeal` is that they can be re-used by future
implementations, whereas putting more into the engine makes it harder to do
this. But path finding may prove too difficult to add to the language.

There is also no "import" keyword in bigdeal still. Standard library modules are
added in only if they are required – i.e. if the user asks for a `grid` and
hasn't currently defined one, the `grid` module is parsed, loaded and added to
the game graph.

I am trying to avoid "import", because I think it both opens up an attack
surface for untrusted user games and introduces all sorts of issues around
import cycles, caching and more resolution. So for now, imports are automatic
and for standard library modules only. There is also the option to add a simple
flag or environment variable that would include additional modules.

With the grid module now defined, it's now possible to successfully parse the
tic-tac-toe example, which is a major milestone!